database created with first entry

+++ db/migrate/20220926063338_create_contacts.rb
@@ -3,7 +3,7 @@ class CreateContacts < ActiveRecord::Migration[6.1]
     create_table :contacts do |t|
       t.string :first_name
       t.string :last_name
-      t.numeric :phone_number
+      t.integer :phone_number
       t.string :email
       t.string :company_name
       t.date :birth_date
